requirements.txt
# SoClose Community Bot — Dependencies
python-telegram-bot==22.6
requests==2.32.5
pandas>=2.0.0,<3
aiosqlite==0.22.1
fastapi==0.128.8
uvicorn==0.40.0
python-dotenv==1.2.1
phonenumbers==9.0.23
twilio>=9.0.0
python-multipart>=0.0.9
streamlit==1.54.0
